feat: wcag compliant text color tiers
This is an early prototype of this feature, but text tiers should be at
least AA compliant. Tier 8 is the text tier for tiers 1-5, tier 9 is the
text tier for tier 7.

At this stage, tier 9 is a little awkward to work with. Each color
variant is actually the same or so close that it's difficult to tell,
because the tier 7 color itself requires a nearly white contrast as it
is.